split dxy,dz,SIP3D out of IDDictTools
