+++ plugins/IDDictTools.cc
@@ -4,7 +4,6 @@
 
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 #include "FWCore/Framework/interface/ConsumesCollector.h"
-#include "FWCore/Utilities/interface/InputTag.h"
 
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/EventSetup.h"
@@ -12,24 +11,16 @@
 #include "cp3_llbb/Framework/interface/Types.h"
 #include "cp3_llbb/TTWAnalysis/interface/DictTool.h"
 
-#include "Helpers.h"
-
 namespace TTWAnalysis {
 /**
  * Additional electron identification variables
  */
-class DictElectronIDVars : public DictTool<pat::Electron>, protected DictPVHelper {
+class DictElectronIDVars : public DictTool<pat::Electron> {
 public:
   DictElectronIDVars(const edm::ParameterSet& config)
     : DictTool<pat::Electron>(config)
-    , DictPVHelper(config)
   {}
   virtual ~DictElectronIDVars() {}
-
-  virtual void doConsumes(const edm::ParameterSet& config, edm::ConsumesCollector&& collector) override
-  {
-    DictPVHelper                ::doConsumes(config, std::forward<edm::ConsumesCollector>(collector));
-  }
 
   virtual Dict evaluate(const pat::Electron& el,
       const edm::Event* event, const edm::EventSetup* /**/,
@@ -37,17 +28,10 @@ public:
   {
     bool elValid{el.originalObjectRef().isNonnull()};
     double eta = elValid ? el.superCluster()->eta() : 0.;
-    const reco::Vertex* pv = event ? getPV(event) : nullptr;
 
     Dict ret;
     ret.add("isEB", el.isEB());
     ret.add("isEE", el.isEE());
-    // Same values used for cut-based electron ID. See:
-    //     https://github.com/cms-sw/cmssw/blob/CMSSW_7_4_15/RecoEgamma/ElectronIdentification/plugins/cuts/GsfEleDzCut.cc#L64
-    //     https://github.com/cms-sw/cmssw/blob/CMSSW_7_4_15/RecoEgamma/ElectronIdentification/plugins/cuts/GsfEleDxyCut.cc#L64
-    ret.add("dxy", pv && elValid ? el.gsfTrack()->dxy(pv->position()) : 0. );
-    ret.add("dz" , pv && elValid ? el.gsfTrack()->dz (pv->position()) : 0. );
-    ret.add("dca", el.dB(pat::Electron::PV3D)/el.edB(pat::Electron::PV3D));
 
     // electron trigger emulation
     //     https://github.com/peruzzim/cmgtools-lite/blob/76X_for2016basis/TTHAnalysis/python/tools/functionsTTH.py#L10-L20
@@ -85,33 +69,20 @@ public:
 /**
  * Additional muon identification variables
  */
-class DictMuonIDVars : public DictTool<pat::Muon>, protected DictPVHelper {
+class DictMuonIDVars : public DictTool<pat::Muon> {
 public:
   DictMuonIDVars(const edm::ParameterSet& config)
     : DictTool<pat::Muon>(config)
-    , DictPVHelper(config)
   {}
   virtual ~DictMuonIDVars() {}
-
-  virtual void doConsumes(const edm::ParameterSet& config, edm::ConsumesCollector&& collector) override
-  {
-    DictPVHelper            ::doConsumes(config, std::forward<edm::ConsumesCollector>(collector));
-  }
 
   virtual Dict evaluate(const pat::Muon& mu,
       const edm::Event* event, const edm::EventSetup* /**/,
       const ProducersManager* /**/, const AnalyzersManager* /**/, const CategoryManager* /**/) const override
   {
     bool muValid{mu.originalObjectRef().isNonnull()};
-    const reco::Vertex* pv = event ? getPV(event) : nullptr;
 
     Dict ret;
-    // Same values used for cut-based muon ID. See:
-    //     https://github.com/cms-sw/cmssw/blob/CMSSW_7_4_15/DataFormats/MuonReco/src/MuonSelectors.cc#L756
-    ret.add("dxy", pv && muValid && mu.muonBestTrack().isNonnull() ? mu.muonBestTrack()->dxy(pv->position()) : 0. );
-    ret.add("dz" , pv && muValid && mu.muonBestTrack().isNonnull() ? mu.muonBestTrack()->dz (pv->position()) : 0. );
-    ret.add("dca", mu.dB(pat::Muon::PV3D)/mu.edB(pat::Muon::PV3D));
-
     ret.add("DPToPT", muValid && mu.innerTrack().isNonnull() ? mu.innerTrack()->ptError()/mu.innerTrack()->pt() : 0);
 
     return ret;
